fix(a2a3): SPMD kernels must not write a shared cache line
The spec_spmd_{producer,consumer} kernels had each SPMD block write one
float into a shared 64B cache line and dcci-flush the whole line. On
silicon each core writes back its full line copy (other slots stale 0),
so the last block to flush clobbers the rest -- false sharing,
last-writer-wins. Passed on sim only (sim models dcci as a no-op);
failed on device with max_diff 8.0. Not a speculative-dispatch bug --
reproduces with the early-resolve flag off.

- Isolate each block on its own cache line (stride 16 floats = 64B):
  kernel_spmd_fill / kernel_spmd_addk write out[block_idx*16];
  kernel_double reads the strided producer output; producer orchestration
  intermediate and consumer golden updated to the strided layout.
- kernel_fill_all: cast i+1 via int32 -- ccec rejects float<-uint32 in
  aicore functions; the test had only ever run from a stale cached .o.
- Document the limitation: aicore-kernel-programming.md "Each block must
  write to its own cache line" and cache-coherency.md "dcci is
  whole-cache-line".

All 3 spec tests now pass on both a2a3sim and a2a3 (device).

+++ tests/st/a2a3/tensormap_and_ringbuffer/spec_spmd_consumer/kernels/aiv/kernel_fill_all.cpp
@@ -49,7 +49,7 @@ extern "C" __aicore__ void kernel_entry(__gm__ int64_t *args) {
 
     uint32_t count = out_tensor->shapes[0];
     for (uint32_t i = 0; i < count; i++) {
-        out[i] = static_cast<float>(i + 1);
+        out[i] = static_cast<float>(static_cast<int32_t>(i + 1));
     }
     dcci(out, ENTIRE_DATA_CACHE, CACHELINE_OUT);
 }
